autocomplete and vendor refactor
